Split commitment tests and commitment contract tests
